slight reorg
