+++ phenex/filters/__init__.py
@@ -2,7 +2,7 @@ from .categorical_filter import CategoricalFilter
 from .codelist_filter import CodelistFilter
 from .relative_time_range_filter import RelativeTimeRangeFilter
 from .value_filter import ValueFilter
-from .date_filter import DateFilter
+from .date_filter import DateFilter, Before, BeforeOrOn, After, AfterOrOn, Date
 from .value import (
     GreaterThan,
     GreaterThanOrEqualTo,
@@ -11,4 +11,3 @@ from .value import (
     EqualTo,
     Value,
 )
-from .date import Before, BeforeOrOn, After, AfterOrOn, Date
